Renaming test to use pytest
