+++ .github/workflows/test.yml
@@ -64,7 +64,7 @@ jobs:
       - name: Run tests
         run: |
           source $VENV
-          poetry run coverage run -m unittest moabb.tests
+          poetry run pytest moabb.tests
           poetry run coverage xml
 
       - name: Run pipelines
