+++ moabb/tests/__init__.py
@@ -2,14 +2,14 @@
 # from moabb.tests.util_tests import *
 import unittest
 
-from moabb.tests.analysis import *
-from moabb.tests.benchmark import *
-from moabb.tests.classification import *
-from moabb.tests.datasets import *
-from moabb.tests.download import *
-from moabb.tests.evaluations import *
-from moabb.tests.paradigms import *
-from moabb.tests.util_tests import TestDownload, TestSetupSeed
+from moabb.tests.test_analysis import *
+from moabb.tests.test_benchmark import *
+from moabb.tests.test_classification import *
+from moabb.tests.test_datasets import *
+from moabb.tests.test_download import *
+from moabb.tests.test_evaluations import *
+from moabb.tests.test_paradigms import *
+from moabb.tests.test_utils import TestDownload, TestSetupSeed
 
 
 if __name__ == "__main__":
