+++ moabb/tests/test_analysis.py
