+++ moabb/tests/test_paradigms.py
